requirements.txt
# 기본 웹 프레임워크
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
python-multipart==0.0.6

# AI/ML 라이브러리 (버전 고정)
torch==2.0.1
transformers==4.30.2
huggingface-hub==0.16.4
sentence-transformers==2.2.2

# 벡터 데이터베이스
chromadb==0.4.15

# 문서 처리
PyPDF2==3.0.1
python-docx==1.1.0

# Google AI
google-generativeai==0.3.2

# 기타 유틸리티
python-dotenv==1.0.0
numpy>=1.24.0
pandas>=2.0.0